Remove loading of obsolete shinyIncubator library
The shinyIncubator library was being used for the progress bar in the shiny reports. This is now in the standard shiny library, and has been removed from shinyIncubator, so this library is no longer relevant (and it is more complicated to install than normal packages because it needs to be installed directly from GitHub using devtools).