Fix tab-separated line parsing bug
For some strange reason the Linux 'cut' command does not recognize tabs when they are passed from the groovy script by bpipe, even though it works fine when run directly from the command line. This was previously solved in the relevant ChIP-seq modules by cutting the 'targets.txt' file lines on spaces. However this is inconsistent with the other pipelines and with the reporting stage of this pipeline, all of which require tab-separated files. Now fixed by converting tabs to spaces on the fly using 'tr'.