Include upset plots as a separated module in the ChIP-Seq pipeline.
The calculation of the combination matrix for an upset plot can be very time intensive and is now moved from the Rmd report to a separated upsetPlot module. The plots prepared in this module are then displayed in the report. All respective config and report files have been updated accordingly.