introduce sub-experiments in DiffBind3 analysis
A new column sub_experiment in contrasts_diffbind.txt is used to summarize the contrasts into sub-experiments. Those sample groups involved in contrasts with the same sub_experiment name are loaded into a diffbind object and are analyzed together. This includes generation of a consensus peakset and normalization of this sample set. The ChIPhelper.diffbind function was updated accordingly.