include new ESSENTIAL_MIN_PEAKLENGTH var in ChIP-Seq pipeline
ESSENTIAL_FRAGLEN is not used anymore as minimum peak lengthin Macs2, but it can be set individually to a smaller value e.g. as recommended for ATAC-Seq. By default, ESSENTIAL_MIN_PEAKLENGTH is not specified which means that Macs2 determines the fragment length automatically and uses it as minimum peak length.